Add cortex-m targets to docs.rs
Since rust-embedded/cortex-m#256 (released in v0.7) the docs links from
cortex-m to this crate link to the target specific version. As these
targets are not included on this crate the links fail.

For example; see the links to RO on this page https://docs.rs/cortex-m/0.7.3/cortex_m/peripheral/cpuid/struct.RegisterBlock.html